Fix getStyle throwing when displayed attribute changes.
See comment in main.js. The issue was due to the updateChart() function deleting the <desc> elemen (which stored the style data returned by getStyle). This caused a null exception when trying to access ".numbers > desc" after changing the attribute via the UI.

This fix uses data- XML attributes on the SVG elements to store our JSON. These data- attributes aren't deleted when .text() is called on the element.